LayoutTests/accessibility/mac/aria-popup-buttons-on-native-elements-expected.txt

Options Options  Options
This tests that native style buttons with aria-haspopup appear as popup buttons.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button.role is 'AXRole: AXPopUpButton'
PASS button.title is 'AXTitle: Options'
PASS button.stringValue is 'AXValue: '
PASS button.role is 'AXRole: AXPopUpButton'
PASS button.title is 'AXTitle: Options'
PASS button.stringValue is 'AXValue: '
PASS button.role is 'AXRole: AXPopUpButton'
PASS button.title is 'AXTitle: Options'
PASS button.stringValue is 'AXValue: '
PASS button.role is 'AXRole: AXPopUpButton'
PASS button.title is 'AXTitle: Options'
PASS button.stringValue is 'AXValue: '
PASS successfullyParsed is true

TEST COMPLETE
